fix: Move sub-pages to a shared folder
