+++ src/PolicyEngine.jsx
@@ -27,7 +27,7 @@ import Header from "./layout/Header";
 import Testimonials from "./pages/Testimonials";
 import CalculatorInterstitial from "./pages/CalculatorInterstitial";
 import CitizensEconomicCouncil from "./applets/CitizensEconomicCouncil";
-import APIDocumentationPage from "./pages/APIDocumentationPage";
+import APIDocumentationPage from "./pages/learn/APIDocumentationPage";
 import SimulationsPage from "./pages/Simulations";
 import CookieConsent from "./modals/CookieConsent";
 import TrafwaCalculator from "./applets/TrafwaCalculator";
@@ -52,11 +52,11 @@ import GiveCalc from "./applets/GiveCalc";
 import { wrappedResponseJson } from "./data/wrappedJson";
 import US2024ElectionCalculator from "./applets/US2024ElectionCalculator";
 import SaltAMTCalculator from "./applets/SaltAMTCalculator";
-import AIPage from "./pages/AI";
-import MicrosimPage from "./pages/MicrosimPage";
-import EducationPage from "./pages/EducationPage";
-import OpenSourcePage from "./pages/OpenSourcePage";
-import BenefitAccessPage from "./pages/BenefitAccessPage";
+import AIPage from "./pages/learn/AI";
+import MicrosimPage from "./pages/learn/MicrosimPage";
+import EducationPage from "./pages/learn/EducationPage";
+import OpenSourcePage from "./pages/learn/OpenSourcePage";
+import BenefitAccessPage from "./pages/learn/BenefitAccessPage";
 
 const PolicyPage = lazy(() => import("./pages/PolicyPage"));
 const HouseholdPage = lazy(() => import("./pages/HouseholdPage"));
